+++ src/runtime/windows.rs
@@ -323,6 +323,9 @@ pub fn run(
                 | ControlCommand::SwayStatus
                 | ControlCommand::SwayEnable
                 | ControlCommand::SwayDisable
+                | ControlCommand::HyprlandStatus
+                | ControlCommand::HyprlandEnable
+                | ControlCommand::HyprlandDisable
                 | ControlCommand::PointerRequest
                 | ControlCommand::PointerCancel => {
                     request.respond(ControlResponse::Err("UNSUPPORTED".into()))